Move some Github related exceptions into cumulusci.core.exceptions and change imports pointing to the old location
