+++ cumulusci/tasks/release_notes/exceptions.py
@@ -1,22 +1,6 @@
 from cumulusci.core.exceptions import CumulusCIException
 
 
-class GithubApiError(CumulusCIException):
-    pass
-
-
-class GithubApiNotFoundError(CumulusCIException):
-    pass
-
-
-class GithubApiNoResultsError(CumulusCIException):
-    pass
-
-
-class GithubApiUnauthorized(CumulusCIException):
-    pass
-
-
 class GithubIssuesError(CumulusCIException):
     pass
