+++ cumulusci/tasks/release_notes/tests/test_parser.py
@@ -8,7 +8,7 @@ from cumulusci.tasks.release_notes.generator import GithubReleaseNotesGenerator
 from cumulusci.tasks.release_notes.parser import ChangeNotesLinesParser
 from cumulusci.tasks.release_notes.parser import GithubIssuesParser
 from cumulusci.tasks.release_notes.parser import IssuesParser
-from cumulusci.tasks.release_notes.exceptions import GithubApiNotFoundError
+from cumulusci.core.exceptions import GithubApiNotFoundError
 from cumulusci.tasks.release_notes.tests.util_github_api import GithubApiTestMixin
 
 PARSER_CONFIG = [
